backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-allauth>=0.61.0
djangorestframework-simplejwt>=5.3.0
dj-rest-auth>=5.0.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
Pillow>=10.0.0
django-filter>=23.0
drf-spectacular>=0.27.0
gunicorn>=21.0.0
whitenoise>=6.0.0
requests>=2.31.0
faker>=20.0.0
factory-boy>=3.3.0
